feat: Przygotowanie pod deploy na Vercel

+++ app/api/postgres/delete-task/route.ts
@@ -1,26 +1,32 @@
-import { executeQuery } from "@/lib/db";
-import { NextResponse } from "next/server";
+import { executeQuery } from '@/lib/db';
+import { NextResponse } from 'next/server';
+import { checkTokenFromRequest } from '../../utils/firebaseAdmin';
 
 export const DELETE = async (req: Request) => {
-    try {
-        const { id } = await req.json();
+  try {
+    const isTokenValid = await checkTokenFromRequest(req);
 
-        const query = `
+    if (!isTokenValid) {
+      return NextResponse.json({ error: 'Invalid token' }, { status: 401 });
+    }
+
+    const { id } = await req.json();
+
+    const query = `
             DELETE FROM tasks
             WHERE id = $1
             RETURNING *;
         `;
-        const values = [id];
+    const values = [id];
 
-        const result = await executeQuery(query, values);
+    const result = await executeQuery(query, values);
 
-        if (result.rowCount === 0) {
-            return NextResponse.json({ error: 'Task not found' }, { status: 404 });
-        }
-
-        return NextResponse.json({ message: 'Task deleted successfully' }, { status: 200 });
-    } catch (error) {
-        console.error('Error deleting task:', error);
-        return NextResponse.json({ error: 'Error deleting task' }, { status: 500 });
+    if (result.rowCount === 0) {
+      return NextResponse.json({ error: 'Task not found' }, { status: 404 });
     }
+
+    return NextResponse.json({ message: 'Task deleted successfully' }, { status: 200 });
+  } catch (error) {
+    return NextResponse.json({ error: 'Error deleting task' }, { status: 500 });
+  }
 };
